refactor: restore previous file structure
