+++ packages/server-renderer/__tests__/renderToString.spec.ts
@@ -11,8 +11,9 @@ import {
   createStaticVNode
 } from 'vue'
 import { escapeHtml, mockWarn } from '@vue/shared'
-import { renderComponent, renderSlot } from '../src/render'
 import { renderToString } from '../src/renderToString'
+import { renderSlot } from '../src/helpers/ssrRenderSlot'
+import { renderComponent } from '../src/helpers/ssrComponent'
 
 mockWarn()
