+++ packages/server-renderer/src/helpers/ssrComponent.ts
@@ -1,0 +1,20 @@
+import {
+  Component,
+  ComponentInternalInstance,
+  createVNode,
+  Slots
+} from '@vue/runtime-core'
+import { Props, renderComponentVNode, SSRBuffer } from '../render'
+import { SSRSlots } from './ssrRenderSlot'
+
+export function renderComponent(
+  comp: Component,
+  props: Props | null = null,
+  children: Slots | SSRSlots | null = null,
+  parentComponent: ComponentInternalInstance | null = null
+): SSRBuffer | Promise<SSRBuffer> {
+  return renderComponentVNode(
+    createVNode(comp, props, children),
+    parentComponent
+  )
+}
